+++ docs/CHANGELOG.md
@@ -93,9 +93,11 @@ Version 1.1.0 (Tamarack): Not yet released
 - Display control binds in spectate mode help test
 - Remove unrar library (no longer needed)
 - Add `cl_swapsgcontrols` command to swap shotgun controls
-- Move settings `alpine_settings.ini`: Big HUD, Linear Pitch, FPGun FOV Scale, swap controls for AR, SG, GN
+- Move settings to `alpine_settings.ini`: Big HUD, linear pitch, FPGun FOV scale, swap controls for AR, SG, GN
 - Add `Alpine Faction` options panel to in-game options menu
 - Add `SkipCutsceneBindAlias` setting to `alpine_settings.ini`
+- Deprecate `IgnoreSwapAssaultRifleControls` and `IgnoreSwapGrenadeControls` in `af_game.tbl` (no longer needed)
+- Move settings to `alpine_settings.ini`: Weapon shake, fullbright characters, lightmaps only
 
 [@natarii](https://github.com/natarii)
 - Add new damage notification hit sound file